+++ tpDSW/domain/alojamiento.js
@@ -1,6 +1,7 @@
 import { DOLAR_USA, Moneda } from "./moneda.js"
 import { Reserva , RangoFechas } from "./reserva.js"
 import { Foto } from "./foto.js"
+import { Direccion } from "./direccionn.js"
 
 export class Alojamiento{
     anfitrion
@@ -23,19 +24,19 @@ export class Alojamiento{
         this.moneda = Moneda.DOLAR_USA; //MONEDA DEFAULT
         this.horarioCheckIn = "";
         this.horarioCheckOut = "";
-        this.direccion = direccion;
+        this.direccion =  new Direccion();  //Lo dejamos vacio se define al momento que hagamos la creación como lo parseamos
         this.cantHuespedesMax = 5; //CANTIDAD DEFAULT QUE SE PUEDE SETEAR
         this.caracteristicas = []
         this.reservas = []
         this.fotos = []
     }
     estasDisponibleEn(rangoDeFechas){
-        this.reservas.some(reserva => rangoFechaEstaEntre(rangoDeFechas, reserva.rangoFechas))
+        this.reservas.some(reserva => this.rangoFechaEstaEntre(rangoDeFechas, reserva.rangoFechas))
     }
 
     rangoFechaEstaEntre(rangoDeFechas1, rangoDeFechas2){
-        fechaInicialCumple = fechaEstaEntre(rangoDeFechas1.fechaInicio, rangoDeFechas2.fechaFin, rangoDeFechas2.fechaFin)
-        fechaFinalCumple = fechaEstaEntre(rangoDeFechas1.fechaFin, rangoDeFechas2.fechaFin, rangoDeFechas2.fechaFin)
+        fechaInicialCumple = this.fechaEstaEntre(rangoDeFechas1.fechaInicio, rangoDeFechas2.fechaFin, rangoDeFechas2.fechaFin)
+        fechaFinalCumple = this.fechaEstaEntre(rangoDeFechas1.fechaFin, rangoDeFechas2.fechaFin, rangoDeFechas2.fechaFin)
         return fechaInicialCumple && fechaFinalCumple
     }
 
@@ -55,34 +56,5 @@ export class Alojamiento{
     getNombre(){return this.nombre}
     getAnfitrion() {
         return this.anfitrion
-    }
-}
-export class Direccion{
-    calle
-    altura
-    ciudad
-    lat
-    long
-
-    constructor(calle, altura, ciudad, lat, long){
-        this.calle = calle;
-        this.altura = altura;
-        this.ciudad = ciudad;
-        this.lat = lat;
-        this.long = long;
-    }
-}
-export class Ciudad{
-    nombre
-    pais
-    constructor(nombre, pais){
-        this.nombre = nombre;
-        this.pais = pais;
-    }
-}
-export class Pais{
-    nombre
-    constructor(nombre){
-        this.nombre = nombre;
     }
 }
