+++ tpDSW/domain/pais.js
@@ -1,0 +1,6 @@
+export class Pais{
+    nombre
+    constructor(nombre){
+        this.nombre = nombre;
+    }
+}
